Remove unchanged dependency
The get() and set() functions of the unchanged package were used to
retrieve and set data from / in objects by using the string object dot
notation (e.g. `'foo[0].bar.baz'`).

The unchanged functions are doing this in a way which prevents the
original object from beeing modified, this behaviour isn't necessary and
even opposing the purpose of this package.

The get() and set() methods were reimplemented in a very simple way
which is sufficient for the scope of this package.

The size of the minified code was shrunk from about 4.5 kb to roughly
1.4 kb (680 bytes gziped).